+++ invilis.halloween/CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to the "halloween" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.1]
+{🟡🔴🔵🟢🟠🟤},
+{⬜🟧🟨}
+### Added
+- Text highlight color (blood red D:)
+- New colors for brackets: VS Code, M&Ms (default), Candy Corn
+- Command for changing bracket palettes
+    - Switch Bracket Palette 'VS Code' | 'M&Ms' | 'Candy Corn'
+### Fixed
+- Screenshots on the marketplace page
+### Removed
+- Herobrine 😱
+### Changed
+- Changelog 😂 (I couldn't help myself)
 
 ## [1.0.1]
 ### Added